requirements-backend.txt
fastapi
uvicorn[standard]
pandas
openpyxl
camelot-py[cv]
pdfplumber
python-multipart
sqlalchemy
pymysql
